pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt>=1.0.0.40.7",
    "dbt-core>=1.10.13",
    "mcp>=1.10.1",
    "pydantic-ai-slim[mcp]>=0.7.2",
    "dbt-postgres>=1.9.1",
    "streamlit>=1.50.0",
    "pydantic-ai>=0.7.2",
    "python-dotenv>=1.1.1",
]
